Move `PrefillFormEffect` to view effect handler
